feat(controlplane): add digester config options

+++ cmd/kafka-sampler/main.go
@@ -114,7 +114,7 @@ func initNeblic(ctx context.Context, logger logging.Logger, config *neblic.Confi
 		options = append(options, sampler.WithTLS())
 	}
 	if config.LimiterOutLimit != 0 {
-		options = append(options, sampler.WithLimiterOutLimit(int64(config.LimiterOutLimit)))
+		options = append(options, sampler.WithLimiterOutLimit(int32(config.LimiterOutLimit)))
 	}
 	if config.UpdateStatsPeriod != 0 {
 		options = append(options, sampler.WithUpdateStatsPeriod(config.UpdateStatsPeriod))
